+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/TextAnalyticsClient.java
@@ -39,7 +39,6 @@ import com.azure.core.util.Context;
 import com.azure.core.util.polling.SyncPoller;
 
 import java.util.Objects;
-import java.util.UUID;
 
 import static com.azure.ai.textanalytics.implementation.Utility.inputDocumentsValidation;
 
@@ -892,8 +891,8 @@ public final class TextAnalyticsClient {
      * {@codesnippet com.azure.ai.textanalytics.TextAnalyticsClient.beginAnalyzeHealthcare#Iterable-RecognizeHealthcareEntityOptions-Context}
      *
      * @param documents A list of {@link TextDocumentInput documents} to be analyzed.
-     * @param options The additional configurable {@link AnalyzeSentimentOptions options} that may be passed when
-     * analyzing sentiments.
+     * @param options The additional configurable {@link RecognizeHealthcareEntityOptions options} that may be passed
+     * when analyzing healthcare task.
      * @param context Additional context that is passed through the Http pipeline during the service call.
      *
      * @return A {@link SyncPoller} that polls the analyze healthcare operation until it has completed, has failed,
@@ -910,23 +909,27 @@ public final class TextAnalyticsClient {
     }
 
     /**
-     * Cancel a long-running operation healthcare task by given job ID.
+     * Cancel a long-running operation healthcare task by given a healthcare task identification number.
      *
      * <p><strong>Code Sample</strong></p>
-     * {@codesnippet com.azure.ai.textanalytics.TextAnalyticsClient.beginCancelAnalyzeHealthcare#UUID-Context}
+     * {@codesnippet com.azure.ai.textanalytics.TextAnalyticsClient.beginCancelAnalyzeHealthcare#String-RecognizeHealthcareEntityOptions-Context}
      *
-     * @param jobId A job identification number.
+     * @param healthcareTaskId The healthcare task identification number.
+     * @param options The additional configurable {@link RecognizeHealthcareEntityOptions options} that may be passed
+     * when cancelling healthcare task.
      * @param context Additional context that is passed through the Http pipeline during the service call.
      *
      * @return A {@link SyncPoller} that polls the analyze healthcare operation until it has completed, has failed,
      * or has been cancelled.
      *
      * @throws TextAnalyticsException If analyze operation fails.
-     * @throws NullPointerException If {@code jobId} is null.
+     * @throws NullPointerException If {@code healthcareTaskId} is null.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
-    public SyncPoller<TextAnalyticsOperationResult, Void> beginCancelAnalyzeHealthcare(UUID jobId, Context context) {
-        return client.analyzeHealthcareAsyncClient.beginCancelAnalyzeHealthcare(jobId, context).getSyncPoller();
+    public SyncPoller<TextAnalyticsOperationResult, Void> beginCancelAnalyzeHealthcare(String healthcareTaskId,
+        RecognizeHealthcareEntityOptions options, Context context) {
+        return client.analyzeHealthcareAsyncClient.beginCancelAnalyzeHealthcare(healthcareTaskId, options, context)
+                   .getSyncPoller();
     }
 
     // Analyze
@@ -941,7 +944,7 @@ public final class TextAnalyticsClient {
      * {@codesnippet com.azure.ai.textanalytics.TextAnalyticsClient.beginAnalyze#Iterable-String-JobManifestTasks-AnalyzeTasksOptions-Context}
      *
      * @param documents A list of {@link TextDocumentInput documents} to be analyzed.
-     * @param displayName Job's display name.
+     * @param displayName The analyze tasks' display name.
      * @param jobManifestTasks A collection of analyze tasks, such as entity recognition, PII, and key phrase tasks.
      * @param options The additional configurable {@link AnalyzeTasksOptions options} that may be passed when
      * analyzing a collection of tasks.