+++ sdk/textanalytics/azure-ai-textanalytics/src/samples/java/com/azure/ai/textanalytics/lro/AnalyzeTasks.java
@@ -49,7 +49,15 @@ public class AnalyzeTasks {
 
         List<TextDocumentInput> documents = Arrays.asList(
             new TextDocumentInput("0", "Elon Musk is the CEO of SpaceX and Tesla.").setLanguage("en"),
-            new TextDocumentInput("1", "My SSN is 859-98-0987").setLanguage("en")
+            new TextDocumentInput("1",
+                "We went to Contoso Steakhouse located at midtown NYC last week for a dinner party, and we adore"
+                    + " the spot! They provide marvelous food and they have a great menu. The chief cook happens to be"
+                    + " the owner (I think his name is John Doe) and he is super nice, coming out of the kitchen and "
+                    + "greeted us all. We enjoyed very much dining in the place! The Sirloin steak I ordered was tender"
+                    + " and juicy, and the place was impeccably clean. You can even pre-order from their online menu at"
+                    + " www.contososteakhouse.com, call 312-555-0176 or send email to order@contososteakhouse.com! The"
+                    + " only complaint I have is the food didn't come fast enough. Overall I highly recommend it!")
+                .setLanguage("en")
         );
 
         EntitiesTask entitiesTask = new EntitiesTask();