+++ sdk/textanalytics/azure-ai-textanalytics/CHANGELOG.md
@@ -1,9 +1,9 @@
 # Release History
 ## 5.1.0-beta.3 (Unreleased)
 **New features**
-- Added support for healthcare analysis feature, it is a long-running operation, and the cancellation supported. 
-- Added support for analyze tasks feature, It analyzes multiple tasks, such as, entity recognition, PII entity recognition 
-and key phrases extraction simultaneously in a list of document.
+- Added support for healthcare analysis feature. It is represented as a long-running operation. Cancellation supported. 
+- Added support for analyze tasks feature, It analyzes multiple tasks (such as, entity recognition, PII entity recognition 
+and key phrases extraction) simultaneously in a list of document.
 
 ## 5.1.0-beta.2 (2020-10-06)
 ### Breaking changes